+++ nbproject/private/private.properties
@@ -1,10 +1,9 @@
-application.args=-test SomeFile.txt AnotherFile.text YetAnotherFile.txt -j System.java Johnson.java Command.java --name "Sean Carrick"
+application.args=-l java -d /home/sean/NetBeansProjects/CLArgParser/src
 compile.on.save=true
 do.depend=true
 do.jar=true
 do.jlink=false
-file.reference.commons-lang3-3.9.jar=C:\\Users\\Integ\\Downloads\\Development\\Libraries\\commons-lang3-3.9-bin\\commons-lang3-3.9\\commons-lang3-3.9.jar
 javac.debug=true
 javadoc.preview=true
 jlink.strip=false
-user.properties.file=C:\\Users\\Integ\\AppData\\Roaming\\NetBeans\\11.1\\build.properties
+user.properties.file=/home/sean/.netbeans/11.1/build.properties
