+++ nbproject/project.properties
@@ -37,11 +37,9 @@ dist.jlink.dir=${dist.dir}/jlink
 dist.jlink.output=${dist.jlink.dir}/ISRCP
 endorsed.classpath=
 excludes=
-file.reference.commons-lang3-3.9.jar=/home/sean/Downloads/Development/commons-lang3-3.9/commons-lang3-3.9.jar
 includes=**
 jar.compress=true
-javac.classpath=\
-    ${file.reference.commons-lang3-3.9.jar}
+javac.classpath=
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=true
