+++ src/com/is2300/rcp/printer/PrintJobWatcher.java
@@ -1,0 +1,99 @@
+/*
+ * Copyright (C) 2019 Integrity Solutions
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.is2300.rcp.printer;
+
+import javax.print.DocPrintJob;
+import javax.print.event.PrintJobAdapter;
+import javax.print.event.PrintJobEvent;
+
+/**
+ *
+ * @author Sean Carrick &lt;sean at carricktrucking dot com&gt;
+ * 
+ * @version 0.1.0
+ * @since 0.1.0
+ */
+public class PrintJobWatcher {
+    //<editor-fold defaultstate="collapsed" desc="Public Static Constants">
+    
+    //</editor-fold>
+
+    //<editor-fold defaultstate="collapsed" desc="Private Member Fields">
+    boolean done;
+    //</editor-fold>
+
+    //<editor-fold defaultstate="collapsed" desc="Static Initializer">
+    static {
+        
+    }
+    //</editor-fold>
+
+    //<editor-fold defaultstate="collapsed" desc="Intstance Initializer">
+    {
+        done = false;
+    }
+    //</editor-fold>
+
+    //<editor-fold defaultstate="collapsed" desc="Constructor(s)">
+    public PrintJobWatcher (DocPrintJob job) {
+        if ( job == null ) {
+            throw new IllegalArgumentException("No job to watch was provided.");
+        }
+        
+        job.addPrintJobListener(new PrintJobAdapter() {
+            public void printJobCanceled(PrintJobEvent e) {
+                allDone();
+            }
+            public void printJobCompleted(PrintJobEvent e) {
+                allDone();
+            }
+            public void printJobNoMoreEvents(PrintJobEvent e) {
+                allDone();
+            }
+            void allDone() {
+                synchronized (PrintJobWatcher.this) {
+                    done = true;
+                    System.out.println("Printing done...");
+                    PrintJobWatcher.this.notify();
+                }
+            }
+        });
+    }
+    //</editor-fold>
+
+    //<editor-fold defaultstate="collapsed" desc="Public Static Methods">
+    
+    //</editor-fold>
+
+    //<editor-fold defaultstate="collapsed" desc="Public Instance Methods">
+    public synchronized void waitForDone() {
+        try {
+            while (!done) {
+                wait();
+            }
+        } catch (InterruptedException e) {
+            System.err.println("Job interupted...");
+        }
+    }
+    //</editor-fold>
+
+    //<editor-fold defaultstate="collapsed" desc="Private Instance Methods">
+    
+    //</editor-fold>
+
+}
